+++ src/ui/Windows/resource.h
@@ -932,6 +932,7 @@
 #define IDC_AC_BUTTON_COPY_TWOFACTORCODE 1615
 #define IDC_AC_STATIC_TWOFACTORCODE     1616
 #define IDC_STATIC_TWOFACTORCODE        1622
+#define IDC_PASSWORD_STRENGTH           1623
 
 // Next default values for new objects
 // 
@@ -939,7 +940,7 @@
 #ifndef APSTUDIO_READONLY_SYMBOLS
 #define _APS_NEXT_RESOURCE_VALUE        573
 #define _APS_NEXT_COMMAND_VALUE         30001
-#define _APS_NEXT_CONTROL_VALUE         1623
+#define _APS_NEXT_CONTROL_VALUE         1624
 #define _APS_NEXT_SYMED_VALUE           557
 #endif
 #endif